requirements.txt
anthropic>=0.40.0
groq
google-genai>=1.0.0
ibm-watsonx-ai>=1.1.0
python-dotenv
pytest>=8.0.0
ijson>=3.2.0
